+++ database/migrations/2023_11_03_140529_create_quotation_details_table.php
@@ -27,12 +27,12 @@ return new class extends Migration
             $table->string('product_code');
 
             $table->integer('quantity');
-            $table->integer('price');
-            $table->integer('unit_price');
-            $table->integer('sub_total');
-            $table->integer('product_discount_amount');
+            $table->decimal('price', 10, 2);
+            $table->decimal('unit_price', 10, 2);
+            $table->decimal('sub_total', 10, 2);
+            $table->decimal('product_discount_amount', 10, 2);
             $table->string('product_discount_type')->default('fixed');
-            $table->integer('product_tax_amount');
+            $table->decimal('product_tax_amount', 10, 2);
 
             $table->timestamps();
         });
